Moved DN to its own file
