+++ hadoop-hdfs-project/hadoop-hdfs-native-client/src/main/native/libhdfspp/lib/connection/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(connection datanodeconnection.cc)
+add_dependencies(connection proto)
